magento graphql-ce#920: Remove name from WishlistOutput
