+++ app/code/Magento/WishlistGraphQl/Model/Resolver/CustomerWishlistResolver.php
@@ -1,0 +1,57 @@
+<?php
+/**
+ * Copyright © Magento, Inc. All rights reserved.
+ * See COPYING.txt for license details.
+ */
+declare(strict_types=1);
+
+namespace Magento\WishlistGraphQl\Model\Resolver;
+
+use Magento\Framework\GraphQl\Config\Element\Field;
+use Magento\Framework\GraphQl\Query\ResolverInterface;
+use Magento\Framework\GraphQl\Schema\Type\ResolveInfo;
+use Magento\Wishlist\Model\WishlistFactory;
+use Magento\Framework\GraphQl\Exception\GraphQlAuthorizationException;
+
+/**
+ * Fetches the Wishlists data according to the GraphQL schema
+ */
+class CustomerWishlistResolver implements ResolverInterface
+{
+    /**
+     * @var WishlistFactory
+     */
+    private $wishlistFactory;
+
+    /**
+     * @param WishlistFactory $wishlistFactory
+     */
+    public function __construct(WishlistFactory $wishlistFactory)
+    {
+        $this->wishlistFactory = $wishlistFactory;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function resolve(
+        Field $field,
+        $context,
+        ResolveInfo $info,
+        array $value = null,
+        array $args = null
+    ) {
+        /* Guest checking */
+        if (false === $context->getExtensionAttributes()->getIsCustomer()) {
+            throw new GraphQlAuthorizationException(__('The current customer isn\'t authorized.'));
+        }
+        $wishlist = $this->wishlistFactory->create()->loadByCustomerId($context->getUserId(), true);
+        return [
+            'id' => (string) $wishlist->getId(),
+            'sharing_code' => $wishlist->getSharingCode(),
+            'updated_at' => $wishlist->getUpdatedAt(),
+            'items_count' => $wishlist->getItemsCount(),
+            'model' => $wishlist,
+        ];
+    }
+}
